requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
pymongo==4.6.1
python-multipart==0.0.6
reportlab==4.0.4
arabic-reshaper==3.0.0
python-bidi==0.4.2